refactor: 패키지 구조 변경 및 RacingGameManager 테스트 에러 해결
