+++ src/main/java/racingcar/domain/RacingGameManager.java
@@ -2,6 +2,9 @@ package racingcar.domain;
 
 import java.util.List;
 
+import racingcar.domain.car.Cars;
+import racingcar.domain.car.car_info.Name;
+
 public class RacingGameManager {
 	private final Cars cars;
 	private final RacingCount racingCount;
