+++ src/main/java/racingcar/domain/car/car_info/RacingNameException.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car.car_info;
 
 public class RacingNameException extends RuntimeException {
 	public static final String LENGTH_MESSAGE = "잘못된 이름입니다. name=";
