+++ src/main/java/racingcar/domain/car/car_info/strategy/NumberStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car.car_info.strategy;
 
 @FunctionalInterface
 public interface NumberStrategy {
