+++ src/main/java/racingcar/domain/car/car_info/strategy/RandomNumberStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car.car_info.strategy;
 
 import java.util.Random;
