+++ src/test/java/racingcar/domain/NameTest.java
@@ -1,7 +1,7 @@
 package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.*;
-import static racingcar.domain.RacingNameException.*;
+import static racingcar.domain.car.car_info.RacingNameException.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,6 +10,9 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import racingcar.domain.car.car_info.Name;
+import racingcar.domain.car.car_info.RacingNameException;
 
 class NameTest {
 	@DisplayName("이름을 생성한다")
