+++ src/test/java/racingcar/domain/RacingGameManagerTest.java
@@ -9,6 +9,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.car.Car;
+import racingcar.domain.car.Cars;
+import racingcar.domain.car.car_info.Name;
+
 class RacingGameManagerTest {
 	private RacingGameManager racingGameManager;
